// frontend/src/shared/api/generated/schemas/guardrailMetadataResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Bab API
 * OpenAPI spec version: 0.1.0
 */

export interface GuardrailMetadataResponse {
  rule_types: string[];
  pii_values: string[];
  phases: string[];
  effects: string[];
  policy_enforcement_modes: string[];
  assignment_enforcement_modes: string[];
  default_rule_effect: string;
  default_rule_phase: string;
  default_rule_priority: number;
  default_policy_enforcement_mode: string;
  default_assignment_enforcement_mode: string;
}